requirements.txt
glfw
PyOpenGL
Pillow
numpy 